requirements.txt
alembic==1.7.7
Babel==2.10.3
blinker==1.5
certifi==2021.5.30
cffi==1.15.1
click==8.0.4
coverage==6.2
cryptography==38.0.1
dataclasses==0.8
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.0.1
ForgeryPy==0.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pbr==5.11.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.5
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
Tempita==0.5.2
typing_extensions==4.1.1
Werkzeug==2.0.3
Whoosh==2.7.4
WTForms==3.0.0
zipp==3.6.0
